Move .js and .jsx components to a new directory
