+++ resources/js/Components/Button.js
@@ -1,17 +1,33 @@
 import React from 'react';
+import './Button.css';
+import { Link } from 'react-router-dom';
 
-export default function Button({ type = 'submit', className = '', processing, children }) {
+const STYLES = ['btn--primary', 'btn--outline', 'btn--test'];
+
+const SIZES = ['btn--medium', 'btn--large'];
+
+export const Button = ({
+                           children,
+                           type,
+                           onClick,
+                           buttonStyle,
+                           buttonSize
+                       }) => {
+    const checkButtonStyle = STYLES.includes(buttonStyle)
+        ? buttonStyle
+        : STYLES[0];
+
+    const checkButtonSize = SIZES.includes(buttonSize) ? buttonSize : SIZES[0];
+
     return (
-        <button
-            type={type}
-            className={
-                `inline-flex items-center px-4 py-2 bg-gray-900 border border-transparent rounded-md font-semibold text-xs text-white uppercase tracking-widest active:bg-gray-900 transition ease-in-out duration-150 ${
-                    processing && 'opacity-25'
-                } ` + className
-            }
-            disabled={processing}
-        >
-            {children}
-        </button>
+        <Link to='/register' className='btn-mobile'>
+            <button
+                className={`btn ${checkButtonStyle} ${checkButtonSize}`}
+                onClick={onClick}
+                type={type}
+            >
+                {children}
+            </button>
+        </Link>
     );
-}
+};
