+++ resources/js/Components/Checkbox.jsx
