+++ resources/js/Components/HomeSection.js
