+++ resources/js/Components/Note.jsx
