+++ resources/js/Components/notes.js
